Changed apt-mirror provider to dynamically generate temp config file.
Previously the provider.conf file had a hardcoded value for the base
folder to download to.  This would ignore any value that was in the
bashellite.conf file or the value passed with the -m option.

This behavior has been corrected to write to a temp file in the /tmp
folder where the "set base_path" is dynically taken from the mirror_tld
and mirror_repo_name passed in variables.