Update/fix provider save location error (#22)
* Fix provider save location

While testing out the reorganization of the gitlab and elastic repos
(using the "__" mechanic), observed that extra directories
were being created erroneously (based on the repo name) and
the correct directory was not being used.

Updated the following providers to correct the issue:
apt-mirror, bandersnatch, google-repo, podman, and reposync.

* Fix podman provider save error

While testing podman save location fix, observed podman provider was
failing to download an image.  The error is in the way the repo_url
parameter was being used by podman provider.  This has been fixed.