Update/refactor some functions (#24)
* Fix time error

While making the change to replace the use of 'fromisoformat' method
for making epoch time from iso time, incorrectly tied the result to the wrong
variable.

* Relocate simple name processing

Seeing as the expected user input, if any is provided, is a list of simple
package names, decided to relocate the simple package name processing
from the processPackageIndex function to the getPackageListFromIndex
function.  Doing this also changed what was being returned by
processPackageIndex, which had to be changed to no longer return a
package name since it now is no longer necessary.  Relevent additional
changes were made as well based on the previous changes.

Renamed function getPackageList to getPackageListFromIndex to be more
precise in indicating what it does.

* Remove global variable use