Add command line option for package name
Added an option to pass a package name for downloading on the command line to
allow working better with gnu parallel.

Add additional error processing

Added returning a empty list to getPackageListFromIndex on error.

Added boolean to return false on error to processPackageIndex and
processPackageFiles.

Correct error handling issues

Updated both downloader scripts to better handle error conditions in such
a way as to be able to output meaningful error messages.